Bid ids is now list of int instead of long

+++ backend/src/main/scala/Auction.scala
@@ -2,7 +2,7 @@ import MyPostgresProfile.api._
 
 case class Auction(auctionId: Long, userId: Int, length: Float, width: Float, height: Float, fragile: Boolean,
                    description: String, departure: java.sql.Timestamp, arrival: java.sql.Timestamp,
-                   auctionEnd: java.sql.Timestamp, startingPrice: Double, bids: List[Long])
+                   auctionEnd: java.sql.Timestamp, startingPrice: Double, bids: List[Int])
 
 class Auctions(tag: Tag) extends Table[Auction](tag, "auctions") {
   def auctionId = column[Long]("auction_id", O.PrimaryKey, O.AutoInc)
@@ -16,7 +16,7 @@ class Auctions(tag: Tag) extends Table[Auction](tag, "auctions") {
   def arrival = column[java.sql.Timestamp]("arrival")
   def auctionEnd = column[java.sql.Timestamp]("auction_end")
   def startingPrice = column[Double]("starting_price")
-  def bids = column[List[Long]]("bid_ids")
+  def bids = column[List[Int]]("bid_ids")
 
   def * = (auctionId, userId, length, width, height, fragile, description, departure, arrival, auctionEnd,
     startingPrice, bids) <> (Auction.tupled, Auction.unapply)
